dvc-cc/pyproject.toml
[tool.poetry]
name = "dvc-cc"
version = "0.8.23"
description = "This connector is used to combine the work of CC (www.curious-containers.cc) and DVC (Open-source Version Control System for Machine Learning Projects)."
authors = ["Jonas Annuscheit <annusch@htw-berlin.de>"]
license = "AGPL-3.0"
readme = "README.md"
repository = "https://github.com/mastaer/dvc-cc.git"
homepage = "https://github.com/mastaer/dvc-cc.git"

[tool.poetry.dependencies]
python = "^3.6"
keyring = "^19.0"
dvc = "^0"
paramiko = "^2.4"
numpy = "^1.16"
cc-faice = "^8.0"
python-gitlab = "^1.8"
pyyaml = "^5.1"
nbconvert = "^5.5"
pandas = "^0"
matplotlib = "^3.1"
seaborn = "^0"
pyrsistent = "^0.15.2"

[tool.poetry.dev-dependencies]

[tool.poetry.scripts]
dvc-cc = 'dvc_cc.main:main'
[build-system]
requires = ["poetry>=0.14.11"]
build-backend = "poetry.masonry.api"